requirements.txt
chardet
coverage
guzzle_sphinx_theme
jyquickhelper
pandas_streaming
pyquickhelper>=1.8
sphinx_rtd_theme
wheel
